Move the click options into the piomatter package
