+++ examples/fbmirror_scaled.py
@@ -41,7 +41,7 @@ import adafruit_blinka_raspberry_pi5_piomatter as piomatter
 import click
 import numpy as np
 import PIL.Image as Image
-import piomatter_click
+import adafruit_blinka_raspberry_pi5_piomatter.click as piomatter_click
 
 with open("/sys/class/graphics/fb0/virtual_size") as f:
     screenx, screeny = [int(word) for word in f.read().split(",")]
